redis

+++ fchat-svr/fchat.status/rpcHandle.go
@@ -25,7 +25,7 @@ func (s *statusSvr) Heartbeat(ctx context.Context, in *rpcPb.Heartbeat2StatusReq
 func (s *statusSvr) SignIn(ctx context.Context, in *rpcPb.SignInReq) (*rpcPb.SignInRsp, error) {
 	flog.Debug("signin uid:%d", in.Uid)
 
-	offsetDay := time.Now().Sub(beginSinginDate).Hours() / 24
+	offsetDay := int(time.Now().Sub(beginSinginDate).Hours() / 24)
 	flog.Debug("offsetDay:%d", offsetDay)
 
 	c := redisCli.Get()
